// Publish/src/types/goods/goods-detail-types.ts
export type AssociationGoodsInfo = {
  goodsNo: string
  saleStatCd: string
  sortSeq: string
  goodsNm: string
  brandNm: string
  contFilePathNm: string
  salePrc: number
  aplyPrc: number
  norPrc: number
  saleRate: number
  buyrAgeLmtCd: string
  iconJson: string[]
}